Moved corrections report into component
